+++ test/integration/gov/epa/emissions/framework/services/cost/MaxEmsRedStrategyTestCase.java
@@ -147,7 +147,7 @@ public class MaxEmsRedStrategyTestCase extends ServiceTestCase {
         return load;
     }
 
-    protected EfficiencyRecord record(Pollutant pollutant, String locale, float efficiency, float cost, int costYear) {
+    protected EfficiencyRecord record(Pollutant pollutant, String locale, float efficiency, double cost, int costYear) {
         EfficiencyRecord record = new EfficiencyRecord();
         record.setPollutant(pollutant);
         record.setLocale(locale);
